Split test_backup in two files
